+++ v2/frontend/wailsjs/go/main/App.d.ts
@@ -22,6 +22,8 @@ export function GetVersion():Promise<string>;
 
 export function QuitApp():Promise<void>;
 
+export function ReorderAccounts(arg1:Array<string>):Promise<void>;
+
 export function SetAutoStart(arg1:boolean):Promise<void>;
 
 export function SetLanguage(arg1:string):Promise<void>;